requirements.txt
# Core AutoGen packages (latest stable)
autogen-agentchat>=0.4.0
autogen-core>=0.4.0
autogen-ext[openai]>=0.4.0
autogen-ext[mcp]>=0.4.0

# LLM and API clients
openai>=1.30.0
python-dotenv>=1.0.0